% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mof_update_calendar.R
\name{get_mof_update_calendar}
\alias{get_mof_update_calendar}
\title{get_mof_update_calendar}
\usage{
get_mof_update_calendar(destdir = NULL)
}
\arguments{
\item{destdir}{output directory}
}
\value{
csv file
}
\description{
get_mof_update_calendar
}
